Adding Schema file+API file; Stable Release 0.0.4
