+++ app.js
@@ -22,7 +22,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/api', apiRouter);
+app.use('/api/getProducts', apiRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
